refactor hex radius+side length u8 -> f32

+++ client/src/map/config.rs
@@ -30,10 +30,7 @@ pub const OIL_BACKGROUND_COLOR: Color = Color {
 pub const HEX_COUNT_SQRT: i16 = 64;
 pub const HEX_COUNT: i16 = HEX_COUNT_SQRT * HEX_COUNT_SQRT;
 pub const HEX_SIDES: u8 = 6;
-pub const HEX_RADIUS: u8 = 32;
-// todo: refactor to f32
-pub const HEX_SIDE_LENGTH: u8 = HEX_RADIUS;
-// todo: same
-pub const HEX_HEIGHT: LazyLock<f32> =
-    LazyLock::new(|| *util::SIN_PI_DIV_3 as f32 * f32::from(HEX_RADIUS) * 2_f32);
+pub const HEX_RADIUS: f32 = 32.;
+pub const HEX_SIDE_LENGTH: f32 = HEX_RADIUS;
+pub const HEX_HEIGHT: LazyLock<f32> = LazyLock::new(|| *util::SIN_PI_DIV_3 as f32 * f32::from(HEX_RADIUS) * 2_f32);
 pub const HEX_ROTATION: f32 = 30_f32;
